% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_dt.R
\name{arrange_dt}
\alias{arrange_dt}
\title{Arrange entries in data.frame}
\usage{
arrange_dt(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{List of variables or name-value pairs of summary/modifications
functions.}
}
\value{
data.table
}
\description{
Analogous function for \code{arrange} in \pkg{dplyr}.
}
\examples{

iris \%>\% arrange_dt(Sepal.Length)

# minus for decreasing order
iris \%>\% arrange_dt(-Sepal.Length)

# arrange by multiple variables
iris \%>\% arrange_dt(Sepal.Length,Petal.Length)

}
\seealso{
\code{\link[dplyr]{arrange}},\code{\link[maditr]{dt_arrange}}
}
