% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cummean.R
\name{cummean}
\alias{cummean}
\title{Cumulative mean}
\usage{
cummean(x)
}
\arguments{
\item{x}{a numeric or complex object,
or an object that can be coerced to one of these.}
}
\description{
Returns a vector whose elements are the cumulative mean of the elements of the argument.
}
\examples{
cummean(1:10)

}
