% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notin.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not in operator}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{vector or NULL}

\item{y}{vector or NULL}
}
\description{
Inverse operation of match.
}
\examples{

"a" \%in\% letters[1:3]
"a" \%notin\% letters[1:3]

1 \%in\% 1:3
1 \%notin\% 1:3
}
