% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/object_size.R
\name{object_size}
\alias{object_size}
\title{Nice printing of report the Space Allocated for an Object}
\usage{
object_size(object)
}
\arguments{
\item{object}{an R object.}
}
\value{
An object of class "object_size"
}
\description{
Provides an estimate of the memory that is being used to store an R object.
A wrapper of `object.size`, but use a nicer printing unit.
}
\examples{

iris \%>\% object_size()

}
