% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_dt.R
\name{pull_dt}
\alias{pull_dt}
\title{Pull out a single variable}
\usage{
pull_dt(data, col)
}
\arguments{
\item{data}{data.frame}

\item{col}{A name of column or index (should be positive).}
}
\value{
vector
}
\description{
Analogous function for \code{pull} in \pkg{dplyr}
}
\examples{
mtcars \%>\% pull_dt(1)
mtcars \%>\% pull_dt(cyl)
mtcars \%>\% pull_dt("cyl")
}
\seealso{
\code{\link[dplyr]{pull}}
}
