% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round0.R
\name{round0}
\alias{round0}
\title{Round a number and make it show zeros}
\usage{
round0(x, digits = 0)
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{Integer indicating the number of decimal places. Defaults to 0.}
}
\value{
A character vector.
}
\description{
Rounds values in its first argument to the specified number of decimal places,
 returning character, ensuring first decimal digits are showed even when they are zeros.
}
\examples{
a = 14.0034
round0(a,2)

b = 10
round0(b,1)

}
\references{
https://stackoverflow.com/questions/42105336/how-to-round-a-number-and-make-it-show-zeros
}
\seealso{
\code{\link[base]{round}}
}
