% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_tibble.R
\name{show_tibble}
\alias{show_tibble}
\title{Use tibble for printing of data.frame}
\usage{
show_tibble(use = FALSE)
}
\arguments{
\item{use}{Logical. Use tibble for printing or not? Default uses \code{FALSE}.}
}
\description{
This function could let the user to choose whether they
would like the output data.table to be printed as the form of data.table
or tibble.
}
\details{
This function actually define or remove a function named as
\code{print.data.table} which let it return the printing of the tibble. One can
 also turn off this mode by \code{rm(print.data.table)}. Note that this procedure
 would not change the form of data.table, but just change the printing output.
}
\examples{

# while the printing form changes, the class of data never changes

iris \%>\% count_dt(Species) -> a
class(a)

show_tibble(TRUE)
a
class(a)

show_tibble(FALSE)
a
class(a)
}
