% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slice_dt.R
\name{slice_dt}
\alias{slice_dt}
\title{Slice rows in data.frame}
\usage{
slice_dt(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{Integer row values.}
}
\value{
data.table
}
\description{
Analogous function for \code{slice} in \pkg{dplyr}
}
\examples{
iris \%>\% slice_dt(1:3)
iris \%>\% slice_dt(1,3)
iris \%>\% slice_dt(c(1,3))
}
\seealso{
\code{\link[dplyr]{slice}}
}
