% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transpose.R
\name{t_dt}
\alias{t_dt}
\title{Efficient transpose of data.frame}
\usage{
t_dt(.data)
}
\arguments{
\item{.data}{A data.frame/data.table/tibble}
}
\value{
A transposed data.frame
}
\description{
An efficient way to transpose data frames(data.frame/data.table/tibble).
}
\details{
This function would return the original data.frame structure,
keeping all the row names and column names. If the row names are not
available or, "V1,V2..." will be provided.
}
\examples{

t_dt(iris)
t_dt(mtcars)
}
