% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dt.R
\name{filter_dt}
\alias{filter_dt}
\title{Filter entries in data.frame}
\usage{
filter_dt(data, ...)
}
\arguments{
\item{data}{data.frame}

\item{...}{List of variables or name-value pairs of summary/modifications
functions.}
}
\value{
data.table
}
\description{
Analogous function for \code{filter} in \pkg{dplyr}.
}
\examples{
iris \%>\% filter_dt(Sepal.Length > 7)
iris \%>\% filter_dt(Sepal.Length > 7,Sepal.Width > 3)
iris \%>\% filter_dt(Sepal.Length > 7 & Sepal.Width > 3)
iris \%>\% filter_dt(Sepal.Length == max(Sepal.Length))
}
\seealso{
\code{\link[dplyr]{filter}}
}
