% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nth.R
\name{nth}
\alias{nth}
\title{Extract the nth value from a vector}
\usage{
nth(v, n = 1)
}
\arguments{
\item{v}{A vector}

\item{n}{A single integer specifying the position. Default uses \code{1}.
Negative integers index from the end
 (i.e. -1L will return the last value in the vector).
 If a double is supplied, it will be silently truncated.}
}
\value{
A single value.
}
\description{
Get the value from a vector with its position.
}
\examples{

x = 1:10
nth(x, 1)
nth(x, 5)
nth(x, -2)

}
