% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_setting.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{\%like\%}
\alias{\%chin\%}
\alias{data.table}
\alias{as.data.table}
\alias{fread}
\alias{fwrite}
\alias{fintersect}
\alias{fsetdiff}
\alias{funion}
\alias{fsetequal}
\alias{frollapply}
\alias{fcoalesce}
\alias{uniqueN}
\alias{rbindlist}
\alias{tables}
\alias{like}
\alias{copy}
\alias{key}
\alias{CJ}
\alias{rleid}
\alias{rleidv}
\alias{fcase}
\alias{between}
\alias{set}
\alias{setDT}
\alias{setDF}
\alias{setnames}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{data.table}{\code{\link[data.table:chmatch]{\%chin\%}}, \code{\link[data.table:like]{\%like\%}}, \code{\link[data.table]{as.data.table}}, \code{\link[data.table]{between}}, \code{\link[data.table:J]{CJ}}, \code{\link[data.table]{copy}}, \code{\link[data.table]{data.table}}, \code{\link[data.table]{fcase}}, \code{\link[data.table:coalesce]{fcoalesce}}, \code{\link[data.table:setops]{fintersect}}, \code{\link[data.table]{fread}}, \code{\link[data.table:froll]{frollapply}}, \code{\link[data.table:setops]{fsetdiff}}, \code{\link[data.table:setops]{fsetequal}}, \code{\link[data.table:setops]{funion}}, \code{\link[data.table]{fwrite}}, \code{\link[data.table:setkey]{key}}, \code{\link[data.table]{like}}, \code{\link[data.table]{rbindlist}}, \code{\link[data.table]{rleid}}, \code{\link[data.table:rleid]{rleidv}}, \code{\link[data.table:assign]{set}}, \code{\link[data.table]{setDF}}, \code{\link[data.table]{setDT}}, \code{\link[data.table:setattr]{setnames}}, \code{\link[data.table]{tables}}, \code{\link[data.table:duplicated]{uniqueN}}}

  \item{stringr}{\code{\link[stringr:pipe]{\%>\%}}}
}}

